Updated source files to build with modified Arduino ESP8266 Core
These changes allow the Duet WiFi Socket Server code to build with a
slightly modified version of the Arduino ESP8266 Core framework. In
particular it can build either with the 2.6.3 release of the core
(currently the latest release) or with a modified version that uses the
latest 3.02 ESP8266 SDK libraries. The fork of the Arduino ESP8266
toolkit that supports this build can be found here:
https://github.com/gloomyandy/Arduino